refactor: Update component import paths after directory restructuring.
