+++ app/my/learning/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import Card from '@/components/common/Card';
-import CancelEnrollmentButton from '@/components/common/My/CancelEnrollmentButton';
+import CancelEnrollmentButton from '@/components/My/CancelEnrollmentButton';
 import { useToast } from '@/components/common/Toast/Context';
 import { ToastType } from '@/components/common/Toast/type';
 import { Lecture } from '@/types/knowledge/lecture';
